gcp/requirements.txt
kubiya
pydantic
google-cloud-compute
google-api-core